requirements.txt
openai
biopython
requests